using context to get app prop in AclSecurityConfiguration
